+++ src/main/java/com/ticketscale/infrastructure/security/SecurityConfigurations.java
@@ -43,6 +43,6 @@ public class SecurityConfigurations {
 
     @Bean
     public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
+        return new Argon2PasswordHasher();
     }
 }